quash warnings on 64-bit

+++ src/alsa.c
@@ -505,7 +505,7 @@ static size_t read_samples(sox_format_t * ft, sox_sample_t *buf, size_t nsamp)
       long n = snd_pcm_readi(alsa->pcm_handle, alsa->buf,
           (nsamp - len)/ft->signal.channels); /* ALSA takes "frame" counts. */
       if (n < 0) {
-        if (xrun_recovery(alsa->pcm_handle, n) < 0) {
+        if (xrun_recovery(alsa->pcm_handle, (int)n) < 0) {
           lsx_fail_errno(ft, SOX_EPERM, "ALSA read error");
           return 0;
         }
